cannon: Add modified sysRead tests for MTCannon
